[lib/base] fix: string -> f32/64 parsing

+++ src/lib/base/strings.cpp
@@ -968,7 +968,7 @@ float string::_float() const {
 // convert a string to a float
 double string::f64() const {
 	bool minus = false;
-	int e = -1;
+	double e = -1;
 	double res = 0;
 	for (int i=0; i<num; i++) {
 		if (e > 0)
@@ -982,7 +982,7 @@ double string::f64() const {
 			if (e < 0)
 				res = res * 10 + (c-48);
 			else
-				res += float(c-48) / (float)e;
+				res += double(c-48) / e;
 		} else if ((c == 'e') or (c == 'E')) {
 			int ex = sub(i+1)._int();
 			res *= pow(10.0, ex);
